ZEN: INTERACTIVE: Tune mgLRU to protect cache used in the last second
Although not identical to the le9 patches that protect a byte-amount of
cache through tunables, multigenerational LRU now supports protecting
cache accessed in the last X milliseconds.

In #218, Yu recommends starting with 1000ms and tuning as needed.  This
looks like a safe default and turning on this feature should help users
that don't know they need it.

Signed-off-by: rmp22 <195054967+rmp22@users.noreply.github.com>
Signed-off-by: Alhaidar Latif <ziyalhaidar@gmail.com>